Move to kitty
... even if I have to see a titlebar

Linux operating systems don't have a terminal emulator with ligature
support which lets you remove the titlebar and retain window
decorations.

A titlebar on a terminal is like letters on the keyboard. It serves no
purpose. (Windows can be moved much faster without having the pointer
anywhere near the titlebar, even on macOS.) Window decorations are
necessary for resizing the window quickly.

iTerm2 is perfect. Linux doesn't have anything close to it.